Latest version of the plugin integrated
Added test keystore

+++ Assets/AppcoinsUnity/Scripts/BashUtils.cs
@@ -9,6 +9,19 @@ using UnityEngine;
 
 public abstract class Terminal
 {
+    protected bool ProcessFailed() {
+        StreamReader reader = new StreamReader(Application.dataPath + "/AppcoinsUnity/Tools/ProcessLog.out");
+        string processLog = reader.ReadToEnd();
+        reader.Close();
+
+        if(processLog.Length != 0)
+        {
+            return true;
+        }
+
+        return false;
+    }
+    
     public ProcessStartInfo InitializeProcessInfo(string terminalPath)
     {
         ProcessStartInfo processInfo = new ProcessStartInfo();
@@ -19,79 +32,22 @@ public abstract class Terminal
         return processInfo;
     }
 
-    public virtual void RunCommand(string cmd, string path) {}
+    public virtual void RunCommand(int buildPhase, string cmd, string cmdArgs, string path, System.Action<int> onDoneCallback) {
+    }
 
-    public void RunTerminalCommand(string terminalPath, string cmd)
+    public void RunTerminalCommand(int buildPhase, string terminalPath, string cmd, string cmdArgs, System.Action<int> onDoneCallback)
     {
-        RunCommand(cmd, "");
+        RunCommand(buildPhase, cmd, cmdArgs, ".", onDoneCallback);
     }
 }
 
 public class Bash : Terminal
 {
-    protected static string TERMINAL_PATH = "/bin/bash";
+    private void RunBashCommand(string terminalPath, int buildPhase, string cmd, string cmdArgs, string path, System.Action<int> onDoneCallback) {
+        CreateSHFileToExecuteCommand(buildPhase, cmd, cmdArgs, path);
 
-    protected virtual void RunBashCommand(string cmd, string path) {}
-
-    public override void RunCommand(string cmd, string path)
-    {
-        Bash t;
-
-        if(Directory.Exists("/Applications/Utilities/Terminal.app") || Directory.Exists("/Applications/Terminal.app"))
-        {
-            t = new BashGUI();
-            t.RunBashCommand(cmd, path);
-        }
-
-        else
-        {
-            t = new BashCommandLine();
-            t.RunBashCommand(cmd, path);
-        }
-    }
-}
-
-public class BashCommandLine : Bash
-{
-    protected override void RunBashCommand(string cmd, string path)
-    {
-        UnityEngine.Debug.Log("Cmd is " + cmd);
-        UnityEngine.Debug.Log("Path is " + path);
-
-        ProcessStartInfo processInfo = InitializeProcessInfo(TERMINAL_PATH);
-        processInfo.RedirectStandardOutput = true;
-        processInfo.RedirectStandardError = true;
-
-        if (path != "")
-        {
-            processInfo.Arguments = "-c \"cd " + path + " && " + cmd + "\"";
-        }
-        else
-        {
-            processInfo.Arguments = "-c \"" + cmd + "\"";
-        }
-
-        UnityEngine.Debug.Log("process args: " + processInfo.Arguments);
-
-        Process newProcess = Process.Start(processInfo);
-
-        string strOutput = newProcess.StandardOutput.ReadToEnd();
-        string strError = newProcess.StandardError.ReadToEnd();
-
-        newProcess.WaitForExit();
-        UnityEngine.Debug.Log(strOutput);
-        UnityEngine.Debug.Log("Process exited with code " + newProcess.ExitCode + "\n and errors: " + strError);
-    }
-}
-
-public class BashGUI : Bash
-{
-    protected override void RunBashCommand(string cmd, string path)
-    {
-        CreateSHFileToExecuteCommand(cmd, path);
-
-        ProcessStartInfo processInfo = InitializeProcessInfo(TERMINAL_PATH);
-        processInfo.FileName = "/Applications/Utilities/Terminal.app/Contents/MacOS/Terminal";
+        ProcessStartInfo processInfo = InitializeProcessInfo(terminalPath);
+        processInfo.FileName = "/bin/bash";
         processInfo.CreateNoWindow = false;
 
 	    processInfo.Arguments = "'" + Application.dataPath + "/AppcoinsUnity/Tools/BashCommand.sh'";
@@ -100,36 +56,99 @@ public class BashGUI : Bash
 	    newProcess.StartInfo = processInfo;
 	    newProcess.Start();
 
+        //TODO FIXME this file might not be created ever!
         //For the process to complete we check with, 5s interval, for the existence of ProcessCompleted.out
-        while(!File.Exists(Application.dataPath + "/AppcoinsUnity/Tools/ProcessCompleted.out"))
-        {
-            Thread.Sleep(5000);
+        bool fileExists;
+        bool condition;
+        do { 
+            fileExists = File.Exists(Application.dataPath + "/AppcoinsUnity/Tools/ProcessCompleted.out");
+            condition = !fileExists;
+            Thread.Sleep(2000);
         }
+        while (condition);
 
         //Now we can safely kill the process
         if(!newProcess.HasExited)
         {
             newProcess.Kill();
         }
+
+        int retCode = ((ProcessFailed() == true) ? -1 : 0);
+        onDoneCallback.Invoke(retCode);
+    }
+
+    public override void RunCommand(int buildPhase, string cmd, string cmdArgs, string path, System.Action<int> onDoneCallback)
+    {
+        string terminalPath = null;
+        int version = -1;
+        int.TryParse(Application.unityVersion.Split('.')[0], out version);
+
+        if(Directory.Exists("/Applications/Utilities/Terminal.app") && version > 5)
+        {
+            terminalPath = "/Applications/Utilities/Terminal.app/Contents/MacOS/Terminal";
+        }
+
+        else
+        {
+            terminalPath = "/bin/bash";
+        }
+
+        RunBashCommand(terminalPath, buildPhase, cmd, cmdArgs, path, onDoneCallback);
     }
 
     //This creates a bash file that gets executed in the specified path
-    private void CreateSHFileToExecuteCommand(string cmd, string path)
+    protected void CreateSHFileToExecuteCommand(int buildPhase, string cmd, string cmdArgs, string path)
     {
+        // Delete all temporary files.
+        if(File.Exists(Application.dataPath + "/AppcoinsUnity/Tools/ProcessCompleted.out"))
+        {
+            File.Delete(Application.dataPath + "/AppcoinsUnity/Tools/ProcessCompleted.out");
+        }
+
+        if(File.Exists(Application.dataPath + "/AppcoinsUnity/Tools/ProcessLog.out"))
+        {
+            File.Delete(Application.dataPath + "/AppcoinsUnity/Tools/ProcessLog.out");
+        }
+
+        if(File.Exists(Application.dataPath + "/AppcoinsUnity/Tools/BashCommand.bat"))
+        {
+            File.Delete(Application.dataPath + "/AppcoinsUnity/Tools/BashCommand.bat");
+        }
+
+        if(File.Exists(Application.dataPath + "/AppcoinsUnity/Tools/BashCommand.sh"))
+        {
+            File.Delete(Application.dataPath + "/AppcoinsUnity/Tools/BashCommand.sh");
+        }       
+
         StreamWriter writer = new StreamWriter(Application.dataPath + "/AppcoinsUnity/Tools/BashCommand.sh", false);
 
         writer.WriteLine("#!/bin/sh");
 
         //Put terminal as first foreground application
+        #if UNITY_5
+        #else
         writer.WriteLine("osascript -e 'activate application \"/Applications/Utilities/Terminal.app\"'");
+        #endif
         writer.WriteLine("cd " + path);
-        writer.WriteLine(cmd);
+        //writer.WriteLine(cmd);
+        if(buildPhase == 2)
+        {
+            writer.WriteLine("if [ \"$(" + cmd + " get-state)\" == \"device\" ]\nthen");
+        }
+
+        // writer.WriteLine(cmd + " " + cmdArgs + " 2> '" + Application.dataPath + "/AppcoinsUnity/Tools/ProcessLog.out' | tee '" + Application.dataPath + "/AppcoinsUnity/Tools/ProcessLog.out'");
+        writer.WriteLine(cmd + " " + cmdArgs + " 2>&1 2>'" + Application.dataPath + "/AppcoinsUnity/Tools/ProcessLog.out'");
+
+        if(buildPhase == 2)
+        {
+            writer.WriteLine("else\necho error: no usb device found > '" + Application.dataPath + "/AppcoinsUnity/Tools/ProcessLog.out'");
+            writer.WriteLine("fi");
+        }
+
         writer.WriteLine("echo 'done' > '" + Application.dataPath + "/AppcoinsUnity/Tools/ProcessCompleted.out'");
         writer.WriteLine("exit");
         // writer.WriteLine("osascript -e 'tell application \"Terminal\" to close first window'");
         writer.Close();
-
-        File.Delete(Application.dataPath + "/AppcoinsUnity/Tools/ProcessCompleted.out");
     }
 }
 
@@ -138,26 +157,90 @@ public class CMD : Terminal
     protected static string TERMINAL_PATH = "cmd.exe";
     private static bool NO_GUI = false;
 
-    public override void RunCommand(string cmd, string path)
+    public override void RunCommand(int buildPhase, string cmd, string cmdArgs, string path, System.Action<int> onDoneCallback)
     {
+        cmd = cmd.Replace("/", "\\");
+        cmd = cmd.Replace("'", "\"");
+        cmdArgs = cmdArgs.Replace("/", "\\");
+        cmdArgs = cmdArgs.Replace("'", "\"");
+        path = path.Replace("/", "\\");
+        path = path.Replace("'", "\"");
+
+        CreateBatchFileToExecuteCommand(buildPhase, cmd, cmdArgs, path);
+
         ProcessStartInfo processInfo = InitializeProcessInfo(TERMINAL_PATH);
         processInfo.CreateNoWindow = NO_GUI;
-        processInfo.RedirectStandardInput = true;
+        processInfo.UseShellExecute = true;
 
-        if (path != "")
-        {
-            processInfo.Arguments = "/c \"cd " + path + " && " + cmd + "\"";
-        }
-        else
-        {
-            processInfo.Arguments = "/c \"" + cmd + "\"";
-        }
+        processInfo.Arguments = "/c '" + Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.bat'";
 
         // Replace string from bash fromat to cmd format
         processInfo.Arguments = processInfo.Arguments.Replace("\"", "");
         processInfo.Arguments = processInfo.Arguments.Replace("'", "\"");
 
         Process newProcess = Process.Start(processInfo);
-        newProcess.WaitForExit();
+
+        bool fileExists;
+        bool condition;
+        do { 
+            fileExists = File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessCompleted.out");
+            condition = !fileExists;
+            Thread.Sleep(2000);
+        }
+        while (condition);
+
+        //Now we can safely kill the process
+        if(!newProcess.HasExited)
+        {
+            newProcess.Kill();
+        }
+
+        int retCode = ((ProcessFailed() == true) ? -1 : 0);
+        onDoneCallback.Invoke(retCode);
+    }
+
+    private void CreateBatchFileToExecuteCommand(int buildPhase, string cmd, string cmdArgs, string path)
+    {
+        // Delete all temporary files.
+        if(File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessCompleted.out"))
+        {
+            File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessCompleted.out");
+        }
+
+        if(File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessLog.out"))
+        {
+            File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessLog.out");
+        }
+
+        if(File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.bat"))
+        {
+            File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.bat");
+        }
+
+        if(File.Exists(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.sh"))
+        {
+            File.Delete(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.sh");
+        }
+
+
+        StreamWriter writer = new StreamWriter(Application.dataPath + "\\AppcoinsUnity\\Tools\\BashCommand.bat", false);
+
+        writer.WriteLine("cd " + path);
+
+        if(buildPhase == 2)
+        {
+            writer.WriteLine("set var=error");
+            writer.WriteLine("for /f \"tokens=*\" %%a in ('" + cmd + " get-state') do set var=%%a");
+            writer.WriteLine("if \"%var%\" == \"device\" (" + cmd + " " + cmdArgs + " 2>\"" + Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessLog.out\")");
+            writer.WriteLine("if \"%var%\" == \"error\" ( echo error >\"" + Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessLog.out\"" + ")");
+        }
+
+        else 
+        {
+            writer.WriteLine("call " + cmd + " " + cmdArgs + " 2>\"" + Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessLog.out\"");
+        }
+
+        writer.WriteLine("echo done >\"" + Application.dataPath + "\\AppcoinsUnity\\Tools\\ProcessCompleted.out\"");
+        writer.Close();
     }
 }
